Consolidate Telegram UX renderer into active telegram path
