+++ projects/polymarket/polyquantbot/reports/forge/telegram_ux_05_interface-path-consolidation.md
@@ -1,0 +1,56 @@
+# Telegram UX 05 — Interface Path Consolidation Into Active Telegram Runtime
+
+Date: 2026-04-22 04:07 (Asia/Jakarta)
+Branch: feature/consolidate-telegram-ui-ux-layer
+Project Root: projects/polymarket/polyquantbot/
+
+## 1. What was built
+
+- Promoted mature Telegram presentation renderer logic into the active Telegram package path by introducing `telegram/view_handler.py` and `telegram/ui_formatter.py` as runtime-owned modules.
+- Rewired active Telegram command/callback runtime imports to consume the new in-package renderer path, removing runtime dependency on `interface/telegram/view_handler.py`.
+- Reduced `interface` Telegram presentation entrypoints into thin compatibility shims so old import paths remain non-breaking while duplicate implementation logic is removed from the interface tree.
+- Archived the deprecated interface Telegram implementation lineage under a dedicated `archive/deprecated/interface/` path with a clear source-of-truth notice.
+
+## 2. Current system architecture (relevant slice)
+
+1. Active Telegram runtime command surfaces under `projects/polymarket/polyquantbot/telegram/` now consume `telegram.view_handler` directly.
+2. `telegram.view_handler` delegates all premium reply rendering to `telegram.ui_formatter` (same mature formatter logic, now in active runtime package).
+3. `interface/ui_formatter.py` and `interface/telegram/view_handler.py` are compatibility-only shims that re-export from active `telegram` modules.
+4. Historical deprecated implementation snapshot and migration notes live under `projects/polymarket/polyquantbot/archive/deprecated/interface/`.
+
+## 3. Files created / modified (full repo-root paths)
+
+- projects/polymarket/polyquantbot/telegram/ui_formatter.py
+- projects/polymarket/polyquantbot/telegram/view_handler.py
+- projects/polymarket/polyquantbot/telegram/command_handler.py
+- projects/polymarket/polyquantbot/telegram/handlers/callback_router.py
+- projects/polymarket/polyquantbot/interface/ui_formatter.py
+- projects/polymarket/polyquantbot/interface/telegram/view_handler.py
+- projects/polymarket/polyquantbot/archive/deprecated/interface/README.md
+- projects/polymarket/polyquantbot/archive/deprecated/interface/telegram_legacy_20260421/ui_formatter.py
+- projects/polymarket/polyquantbot/archive/deprecated/interface/telegram_legacy_20260421/view_handler.py
+- projects/polymarket/polyquantbot/reports/forge/telegram_ux_05_interface-path-consolidation.md
+- PROJECT_STATE.md
+
+## 4. What is working
+
+- Active Telegram runtime now resolves mature presentation rendering through `projects/polymarket/polyquantbot/telegram/*` package-local imports.
+- Legacy interface renderer import paths remain functional through shim exports (compatibility without duplicate logic ownership).
+- Deprecated Telegram interface implementation is no longer an active truth path and is explicitly archived for traceability.
+- `/start`, `/help`, `/status`, and unknown-command presentation foundations continue to resolve through one coherent renderer layer with explicit paper-only/public-safe posture inherited from existing formatter content.
+
+## 5. Known issues
+
+- Interface package still contains non-Telegram view modules used by other runtime surfaces (outside this task scope), so full interface package retirement is not yet possible.
+- Compatibility shims are intentionally retained; callers should be migrated gradually to `projects.polymarket.polyquantbot.telegram.*` imports.
+
+## 6. What is next
+
+- Continue COMMANDER review for this STANDARD lane and decide timeline for removing compatibility shims once remaining import sites are migrated.
+- Keep Telegram UX consolidation follow-up focused on onboarding repetition refinement only (no runtime architecture expansion).
+
+Validation Tier   : STANDARD
+Claim Level       : NARROW INTEGRATION
+Validation Target : Consolidate mature Telegram renderer ownership under active `projects/polymarket/polyquantbot/telegram` path while preserving runtime behavior and compatibility.
+Not in Scope      : Fly deploy changes, wallet lifecycle, portfolio architecture rewrite, live-trading enablement, production-capital readiness claims.
+Suggested Next    : COMMANDER review
